+++ Tests/XcodeGenKitTests/ProjectSpecTests.swift
@@ -269,9 +269,9 @@ class ProjectSpecTests: XCTestCase {
                                                     productName: "App",
                                                     deploymentTarget: Version(major: 0, minor: 1, patch: 2),
                                                     settings: Settings(buildSettings: ["foo": "bar"],
-                                                                       configSettings: ["foo" : Settings(buildSettings: ["nested": "config"],
-                                                                                                         configSettings: [:],
-                                                                                                         groups: ["config-setting-group"])],
+                                                                       configSettings: ["foo": Settings(buildSettings: ["nested": "config"],
+                                                                                                        configSettings: [:],
+                                                                                                        groups: ["config-setting-group"])],
                                                                        groups: ["setting-group"]),
                                                     configFiles: ["foo": "bar"],
                                                     sources: [TargetSource(path: "Source",
@@ -331,12 +331,13 @@ class ProjectSpecTests: XCTestCase {
                                                                            action: .compilerSpec("com.apple.xcode.tools.swift.compiler"),
                                                                            name: nil,
                                                                            outputFiles: ["bar"],
-                                                                           outputFilesCompilerFlags: ["foo"]),],
+                                                                           outputFilesCompilerFlags: ["foo"])],
                                                     scheme: TargetScheme(testTargets: [Scheme.Test.TestTarget(name: "test target",
                                                                                                               randomExecutionOrder: false,
                                                                                                               parallelizable: false)],
                                                                          configVariants: ["foo"],
                                                                          gatherCoverageData: true,
+                                                                         disableMainThreadChecker: true,
                                                                          commandLineArguments: ["foo": true],
                                                                          environmentVariables: [XCScheme.EnvironmentVariable(variable: "environmentVariable",
                                                                                                                              value: "bar",
@@ -374,6 +375,7 @@ class ProjectSpecTests: XCTestCase {
                                                                                                                                 parallelizable: false)],
                                                                                            configVariants: ["foo"],
                                                                                            gatherCoverageData: true,
+                                                                                           disableMainThreadChecker: true,
                                                                                            commandLineArguments: ["foo": true],
                                                                                            environmentVariables: [XCScheme.EnvironmentVariable(variable: "environmentVariable",
                                                                                                                                                value: "bar",
@@ -386,14 +388,14 @@ class ProjectSpecTests: XCTestCase {
                                                                                                                                 settingsTarget: "foo")]),
                                                                       attributes: ["foo": "bar"])],
                                    settings: Settings(buildSettings: ["foo": "bar"],
-                                                      configSettings: ["foo" : Settings(buildSettings: ["nested": "config"],
-                                                                                        configSettings: [:],
-                                                                                        groups: ["config-setting-group"])],
+                                                      configSettings: ["foo": Settings(buildSettings: ["nested": "config"],
+                                                                                       configSettings: [:],
+                                                                                       groups: ["config-setting-group"])],
                                                       groups: ["setting-group"]),
                                    settingGroups: ["foo": Settings(buildSettings: ["foo": "bar"],
-                                                                   configSettings: ["foo" : Settings(buildSettings: ["nested": "config"],
-                                                                                                     configSettings: [:],
-                                                                                                     groups: ["config-setting-group"])],
+                                                                   configSettings: ["foo": Settings(buildSettings: ["nested": "config"],
+                                                                                                    configSettings: [:],
+                                                                                                    groups: ["config-setting-group"])],
                                                                    groups: ["setting-group"])],
                                    schemes: [Scheme(name: "scheme",
                                                     build: Scheme.Build(targets: [Scheme.BuildTarget(target: "foo",
@@ -419,6 +421,7 @@ class ProjectSpecTests: XCTestCase {
                                                                                                                         enabled: false)]),
                                                     test: Scheme.Test(config: "Config",
                                                                       gatherCoverageData: true,
+                                                                      disableMainThreadChecker: true,
                                                                       randomExecutionOrder: false,
                                                                       parallelizable: false,
                                                                       commandLineArguments: ["foo": true],
@@ -500,7 +503,7 @@ class ProjectSpecTests: XCTestCase {
     }
 }
 
-fileprivate func expectValidationError(_ project: Project, _ expectedError: SpecValidationError.ValidationError, file: String = #file, line: Int = #line) throws {
+private func expectValidationError(_ project: Project, _ expectedError: SpecValidationError.ValidationError, file: String = #file, line: Int = #line) throws {
     do {
         try project.validate()
     } catch let error as SpecValidationError {